+++ src/main/java/com/cibertec/air/dao/AeropuertoDao.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.cibertec.air.domain.Aeropuerto;
 
 @Repository
-public interface AeropuertoDao extends JpaRepository<Aeropuerto, Integer>{
+public interface AeropuertoDao extends JpaRepository<Aeropuerto, Long>{
 
 }
